+++ src/views/DEX/components/MyOrderHistory/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./MyOrderHistory";
